Porting Guide and Test Data
Add `PORTING.md` with guidelines on how to port the fts_fuzzy_match to
new languages and test the results using a predefined test strategy.

Add a test file for the original JavaScript v0.1.0 algorithm, and copy
the result data to `fts_fuzzy_match/0.1.0/expected_results.txt` so it
can be used to validate ports of the algorithms.

(solves some pending tasks from #16)