C++ v0.1.0 Test Data
Implement test for the C++ implementation of v0.1.0 of the algorithm,
which now becomes the official implementation reference for all tests
data, so its `test_results.txt` is copied to:

    fts_fuzzy_match/0.1.0/expected_results.txt

(C++ test and reference data for v0.2.0 will be added soon).

Improve the tests for the JavaScript v0.1.0 implementation, comparing
its results to the `0.1.0/expected_results.txt`, just to confirm that
both original algorithms behave identically, and to provide a full
example of how to implement and run tests.

(further solves some pending tasks from #16)