Add JS v0.2.0 Test
Add test code to the JavaScript v0.2.0 implementation by @nrgwsth.
The test passed 100%.

In `fts_fuzzy_match/REAMDE.md` tweak the Implementations table, adding
a new "test" column indicating the test status of each algorithm, or
the lack thereof; the "source file" was dropped, and links to the
specific implementation folder and source file were moved to the "lang"
and "version" columns instead.